requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==17.0.0
bitsandbytes==0.49.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.4.2
cycler==0.12.1
datasets==3.6.0
decord==0.6.0
diffusers==0.37.0
dill==0.3.8
exceptiongroup==1.3.1
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
Jinja2==3.1.6
kiwisolver==1.5.0
kornia==0.8.2
kornia_rs==0.1.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
peft==0.18.1
pillow==12.1.1
platformdirs==4.9.4
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
ruff==0.9.10
safetensors==0.7.0
sentencepiece==0.2.1
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sympy==1.14.0
tokenizers==0.21.4
torch==2.10.0
torchao==0.16.0
torchdata==0.11.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.55.4
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.1
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
